% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbWebForms.R
\name{dbTypeToHTMLInputType}
\alias{dbTypeToHTMLInputType}
\title{Convert SQL data types to likely HTML input types}
\usage{
dbTypeToHTMLInputType(db_type)
}
\arguments{
\item{db_type}{A string, the SQL data type to convert to HTML input type.}
}
\value{
A character vector, the HTML input type and attributes likely associated with the SQL data type.
}
\description{
Convert SQL data types to likely HTML input types
}
\examples{
dbTypeToHTMLInputType("int")
}
